Enhancement: Rename SlowTestReporter to Reporter\Reporter
